test: group app-layout responsive tests by part
Rename content-offset.test.js to responsive.test.js and nest initial
render, slot change and resize under drawer / touch-optimized: false /
touch-optimized: true. Add resize tests for overlay and touch-optimized
mode changes driven by the CSS properties.
